feat: add FIPS operator bundle checks as informative tests
with fips-operator-bundle-check and fips-operator-bundle-check-ta
being added as required tasks, they may fail even if the operator
bundle is FIPS compliant for a specific OCP version. This is because
on a bundle level, there's no way to determine the target OCP version
for it and hence, we cannot apply the appropriate exception list during
the check. This can only be determined when dealing with an FBC fragment.
Hence, this commit adds the operator bundle check as informative.

Signed-off-by: Yashvardhan Nanavati <yashn@bu.edu>